Moves algorithms to separate module
